+++ src/include/cdb/cdbfts.h
@@ -73,7 +73,7 @@ extern void FtsShmemInit(void);
 
 extern bool FtsTestConnection(CdbComponentDatabaseInfo *db_to_test, bool full_scan);
 extern void FtsReConfigureMPP(bool create_new_gangs);
-extern void FtsHandleNetFailure(SegmentDatabaseDescriptor **, int);
+extern void FtsHandleNetFailure(SegmentDatabaseDescriptor **, int64);
 extern bool FtsTestSegmentDBIsDown(SegmentDatabaseDescriptor *, int);
 
 extern bool verifyFtsSyncCount(void);
